server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.4.2
python-dotenv>=1.0.0
email-validator>=2.1.0.post1